poc/agent-weather/requirements.txt
# Core dependencies
langgraph==0.0.30
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
